@@ -1,39 +1,40 @@
 import 'dart:typed_data';
-import 'package:meta/meta.dart';
+
 import 'package:bip32/src/utils/ecurve.dart' show isPoint;
 import 'package:bs58check/bs58check.dart' as bs58check;
 
 import '../crypto.dart';
 import '../models/networks.dart';
 import '../payments/index.dart' show PaymentData;
-import '../utils/script.dart' as bscript;
 import '../utils/constants/op.dart';
+import '../utils/script.dart' as bscript;
 
 class P2PKH {
   PaymentData data;
   NetworkType network;
-  P2PKH({@required data, network}) {
-    this.network = network ?? bitcoin;
-    this.data = data;
+
+  P2PKH({required this.data, NetworkType? network})
+      : network = network ?? bitcoin {
     _init();
   }
+
   _init() {
     if (data.address != null) {
-      _getDataFromAddress(data.address);
+      _getDataFromAddress(data.address!);
       _getDataFromHash();
     } else if (data.hash != null) {
       _getDataFromHash();
     } else if (data.output != null) {
-      if (!isValidOutput(data.output))
+      if (!isValidOutput(data.output!))
         throw new ArgumentError('Output is invalid');
-      data.hash = data.output.sublist(3, 23);
+      data.hash = data.output!.sublist(3, 23);
       _getDataFromHash();
     } else if (data.pubkey != null) {
-      data.hash = hash160(data.pubkey);
+      data.hash = hash160(data.pubkey!);
       _getDataFromHash();
       _getDataFromChunk();
     } else if (data.input != null) {
-      List<dynamic> _chunks = bscript.decompile(data.input);
+      List<dynamic> _chunks = bscript.decompile(data.input)!;
       _getDataFromChunk(_chunks);
       if (_chunks.length != 2) throw new ArgumentError('Input is invalid');
       if (!bscript.isCanonicalScriptSignature(_chunks[0]))
@@ -45,12 +46,12 @@ class P2PKH {
     }
   }
 
-  void _getDataFromChunk([List<dynamic> _chunks]) {
+  void _getDataFromChunk([List<dynamic>? _chunks]) {
     if (data.pubkey == null && _chunks != null) {
       data.pubkey = (_chunks[1] is int)
           ? new Uint8List.fromList([_chunks[1]])
           : _chunks[1];
-      data.hash = hash160(data.pubkey);
+      data.hash = hash160(data.pubkey!);
       _getDataFromHash();
     }
     if (data.signature == null && _chunks != null)
@@ -58,7 +59,7 @@ class P2PKH {
           ? new Uint8List.fromList([_chunks[0]])
           : _chunks[0];
     if (data.input == null && data.pubkey != null && data.signature != null) {
-      data.input = bscript.compile([data.signature, data.pubkey]);
+      data.input = bscript.compile([data.signature!, data.pubkey!]);
     }
   }
 
@@ -66,7 +67,7 @@ class P2PKH {
     if (data.address == null) {
       final payload = new Uint8List(21);
       payload.buffer.asByteData().setUint8(0, network.pubKeyHash);
-      payload.setRange(1, payload.length, data.hash);
+      payload.setRange(1, payload.length, data.hash!);
       data.address = bs58check.encode(payload);
     }
     if (data.output == null) {
@@ -86,7 +87,7 @@ class P2PKH {
     if (version != network.pubKeyHash)
       throw new ArgumentError('Invalid version or Network mismatch');
     data.hash = payload.sublist(1);
-    if (data.hash.length != 20) throw new ArgumentError('Invalid address');
+    if (data.hash!.length != 20) throw new ArgumentError('Invalid address');
   }
 }
 
