@@ -1,14 +1,16 @@
 import 'dart:typed_data';
+
 import 'package:hex/hex.dart';
-import 'payments/index.dart' show PaymentData;
-import 'payments/p2pkh.dart' show P2PKH;
-import 'payments/p2pk.dart' show P2PK;
-import 'payments/p2wpkh.dart' show P2WPKH;
-import 'crypto.dart' as bcrypto;
+
 import 'classify.dart';
+import 'crypto.dart' as bcrypto;
+import 'payments/index.dart' show PaymentData;
+import 'payments/p2pk.dart' show P2PK;
+import 'payments/p2pkh.dart' show P2PKH;
+import 'payments/p2wpkh.dart' show P2WPKH;
 import 'utils/check_types.dart';
-import 'utils/script.dart' as bscript;
 import 'utils/constants/op.dart';
+import 'utils/script.dart' as bscript;
 import 'utils/varuint.dart' as varuint;
 
 const DEFAULT_SEQUENCE = 0xffffffff;
@@ -19,23 +21,25 @@ const SIGHASH_ANYONECANPAY = 0x80;
 const ADVANCED_TRANSACTION_MARKER = 0x00;
 const ADVANCED_TRANSACTION_FLAG = 0x01;
 final EMPTY_SCRIPT = Uint8List.fromList([]);
-final EMPTY_WITNESS = new List<Uint8List>();
+final List<Uint8List> EMPTY_WITNESS = [];
 final ZERO = HEX
     .decode('0000000000000000000000000000000000000000000000000000000000000000');
 final ONE = HEX
     .decode('0000000000000000000000000000000000000000000000000000000000000001');
 final VALUE_UINT64_MAX = HEX.decode('ffffffffffffffff');
-final BLANK_OUTPUT =
-    new Output(script: EMPTY_SCRIPT, valueBuffer: VALUE_UINT64_MAX);
+final BLANK_OUTPUT = new Output(
+    script: EMPTY_SCRIPT, valueBuffer: VALUE_UINT64_MAX as Uint8List);
 
 class Transaction {
   int version = 1;
   int locktime = 0;
   List<Input> ins = [];
   List<Output> outs = [];
+
   Transaction();
 
-  int addInput(Uint8List hash, int index, [int sequence, Uint8List scriptSig]) {
+  int addInput(Uint8List hash, int index,
+      [int? sequence, Uint8List? scriptSig]) {
     ins.add(new Input(
         hash: hash,
         index: index,
@@ -51,17 +55,16 @@ class Transaction {
   }
 
   bool hasWitnesses() {
-    var witness = ins.firstWhere(
-        (input) => input.witness != null && input.witness.length != 0,
-        orElse: () => null);
-    return witness != null;
+    var witness = ins.indexWhere(
+        (input) => input.witness != null && input.witness!.length != 0);
+    return witness != -1;
   }
 
   setInputScript(int index, Uint8List scriptSig) {
     ins[index].script = scriptSig;
   }
 
-  setWitness(int index, List<Uint8List> witness) {
+  setWitness(int index, List<Uint8List>? witness) {
     ins[index].witness = witness;
   }
 
@@ -77,8 +80,8 @@ class Transaction {
     var hashSequence = ZERO;
 
     writeSlice(slice) {
-      tbuffer.setRange(toffset, toffset + slice.length, slice);
-      toffset += slice.length;
+      tbuffer.setRange(toffset, toffset + slice.length as int, slice);
+      toffset += slice.length as int;
     }
 
     writeUInt8(i) {
@@ -144,8 +147,8 @@ class Transaction {
 
     if ((hashType & 0x1f) != SIGHASH_SINGLE &&
         (hashType & 0x1f) != SIGHASH_NONE) {
-      var txOutsSize =
-          outs.fold(0, (sum, output) => sum + 8 + varSliceSize(output.script));
+      var txOutsSize = outs.fold(
+          0, (int sum, output) => sum + 8 + varSliceSize(output.script!));
       tbuffer = new Uint8List(txOutsSize);
       bytes = tbuffer.buffer.asByteData();
       toffset = 0;
@@ -157,7 +160,7 @@ class Transaction {
     } else if ((hashType & 0x1f) == SIGHASH_SINGLE && inIndex < outs.length) {
       // SIGHASH_SINGLE only hash that according output
       var output = outs[inIndex];
-      tbuffer = new Uint8List(8 + varSliceSize(output.script));
+      tbuffer = new Uint8List(8 + varSliceSize(output.script!));
       bytes = tbuffer.buffer.asByteData();
       toffset = 0;
       writeUInt64(output.value);
@@ -188,7 +191,7 @@ class Transaction {
     if (inIndex >= ins.length) return ONE;
     // ignore OP_CODESEPARATOR
     final ourScript =
-        bscript.compile(bscript.decompile(prevOutScript).where((x) {
+        bscript.compile(bscript.decompile(prevOutScript)!.where((x) {
       return x != OPS['OP_CODESEPARATOR'];
     }).toList());
     final txTmp = Transaction.clone(this);
@@ -248,10 +251,10 @@ class Transaction {
     return (hasWitness ? 10 : 8) +
         varuint.encodingLength(ins.length) +
         varuint.encodingLength(outs.length) +
-        ins.fold(0, (sum, input) => sum + 40 + varSliceSize(input.script)) +
-        outs.fold(0, (sum, output) => sum + 8 + varSliceSize(output.script)) +
+        ins.fold(0, (sum, input) => sum + 40 + varSliceSize(input.script!)) +
+        outs.fold(0, (sum, output) => sum + 8 + varSliceSize(output.script!)) +
         (hasWitness
-            ? ins.fold(0, (sum, input) => sum + vectorSize(input.witness))
+            ? ins.fold(0, (sum, input) => sum + vectorSize(input.witness!))
             : 0);
   }
 
@@ -275,7 +278,7 @@ class Transaction {
     return (weight() / 4).ceil();
   }
 
-  Uint8List toBuffer([Uint8List buffer, int initialOffset]) {
+  Uint8List toBuffer([Uint8List? buffer, int? initialOffset]) {
     return this._toBuffer(buffer, initialOffset, true);
   }
 
@@ -304,7 +307,7 @@ class Transaction {
     return HEX.encode(getHash().reversed.toList());
   }
 
-  _toBuffer([Uint8List buffer, initialOffset, bool _ALLOW_WITNESS = false]) {
+  _toBuffer([Uint8List? buffer, initialOffset, bool _ALLOW_WITNESS = false]) {
     // _ALLOW_WITNESS is used to separate witness part when calculating tx id
     if (buffer == null) buffer = new Uint8List(_byteLength(_ALLOW_WITNESS));
 
@@ -314,7 +317,7 @@ class Transaction {
     var offset = initialOffset ?? 0;
 
     writeSlice(slice) {
-      buffer.setRange(offset, offset + slice.length, slice);
+      buffer!.setRange(offset, offset + slice.length, slice);
       offset += slice.length;
     }
 
@@ -339,7 +342,7 @@ class Transaction {
     }
 
     writeVarInt(i) {
-      varuint.encode(i, buffer, offset);
+      varuint.encode(i, buffer!, offset);
       offset += varuint.encodingLength(i);
     }
 
@@ -444,7 +447,7 @@ class Transaction {
       return i;
     }
 
-    Uint8List readSlice(n) {
+    Uint8List readSlice(int n) {
       offset += n;
       return buffer.sublist(offset - n, offset);
     }
@@ -517,35 +520,37 @@ class Transaction {
     bool noStrict = false,
   }) {
     return Transaction.fromBuffer(
-      HEX.decode(hex),
+      HEX.decode(hex) as Uint8List,
       noStrict: noStrict,
     );
   }
 
   @override
   String toString() {
+    var buf = StringBuffer();
     this.ins.forEach((txInput) {
-      print(txInput.toString());
+      buf.write(txInput.toString());
     });
     this.outs.forEach((txOutput) {
-      print(txOutput.toString());
+      buf.write(txOutput.toString());
     });
+    return buf.toString();
   }
 }
 
 class Input {
-  Uint8List hash;
-  int index;
-  int sequence;
-  int value;
-  Uint8List script;
-  Uint8List signScript;
-  Uint8List prevOutScript;
-  String prevOutType;
-  bool hasWitness;
-  List<Uint8List> pubkeys;
-  List<Uint8List> signatures;
-  List<Uint8List> witness;
+  Uint8List? hash;
+  int? index;
+  int? sequence;
+  int? value;
+  Uint8List? script;
+  Uint8List? signScript;
+  Uint8List? prevOutScript;
+  String? prevOutType;
+  bool hasWitness = false;
+  List<Uint8List?>? pubkeys;
+  List<Uint8List?>? signatures;
+  List<Uint8List>? witness;
 
   Input(
       {this.hash,
@@ -558,19 +563,18 @@ class Input {
       this.signatures,
       this.witness,
       this.prevOutType}) {
-    this.hasWitness = false; // Default value
-    if (this.hash != null && !isHash256bit(this.hash))
+    if (this.hash != null && !isHash256bit(this.hash!))
       throw new ArgumentError('Invalid input hash');
-    if (this.index != null && !isUint(this.index, 32))
+    if (this.index != null && !isUint(this.index!, 32))
       throw new ArgumentError('Invalid input index');
-    if (this.sequence != null && !isUint(this.sequence, 32))
+    if (this.sequence != null && !isUint(this.sequence!, 32))
       throw new ArgumentError('Invalid input sequence');
-    if (this.value != null && !isShatoshi(this.value))
+    if (this.value != null && !isShatoshi(this.value!))
       throw ArgumentError('Invalid ouput value');
   }
 
   factory Input.expandInput(Uint8List scriptSig, List<Uint8List> witness,
-      [String type, Uint8List scriptPubKey]) {
+      [String? type, Uint8List? scriptPubKey]) {
     if (type == null || type == '') {
       var ssType = classifyInput(scriptSig);
       var wsType = classifyWitness(witness);
@@ -579,45 +583,50 @@ class Input {
       type = ssType ?? wsType;
     }
     if (type == SCRIPT_TYPES['P2WPKH']) {
-      P2WPKH p2wpkh = new P2WPKH(data: new PaymentData(witness: witness));
+      P2WPKH p2wpkh = P2WPKH(data: new PaymentData(witness: witness));
       return new Input(
           prevOutScript: p2wpkh.data.output,
           prevOutType: SCRIPT_TYPES['P2WPKH'],
-          pubkeys: [p2wpkh.data.pubkey],
-          signatures: [p2wpkh.data.signature]);
+          pubkeys: [p2wpkh.data.pubkey!],
+          signatures: [p2wpkh.data.signature!]);
     } else if (type == SCRIPT_TYPES['P2PKH']) {
       P2PKH p2pkh = new P2PKH(data: new PaymentData(input: scriptSig));
       return new Input(
           prevOutScript: p2pkh.data.output,
           prevOutType: SCRIPT_TYPES['P2PKH'],
-          pubkeys: [p2pkh.data.pubkey],
-          signatures: [p2pkh.data.signature]);
+          pubkeys: [p2pkh.data.pubkey!],
+          signatures: [p2pkh.data.signature!]);
     } else if (type == SCRIPT_TYPES['P2PK']) {
       P2PK p2pk = new P2PK(data: new PaymentData(input: scriptSig));
       return new Input(
           prevOutType: SCRIPT_TYPES['P2PK'],
           pubkeys: [],
-          signatures: [p2pk.data.signature]);
+          signatures: [p2pk.data.signature!]);
     }
+    throw UnsupportedError('type "$type"');
   }
 
   factory Input.clone(Input input) {
     return new Input(
-      hash: input.hash != null ? Uint8List.fromList(input.hash) : null,
+      hash: input.hash != null ? Uint8List.fromList(input.hash!) : null,
       index: input.index,
-      script: input.script != null ? Uint8List.fromList(input.script) : null,
+      script: input.script != null ? Uint8List.fromList(input.script!) : null,
       sequence: input.sequence,
       value: input.value,
       prevOutScript: input.prevOutScript != null
-          ? Uint8List.fromList(input.prevOutScript)
+          ? Uint8List.fromList(input.prevOutScript!)
           : null,
       pubkeys: input.pubkeys != null
-          ? input.pubkeys.map(
-              (pubkey) => pubkey != null ? Uint8List.fromList(pubkey) : null)
+          ? input.pubkeys!
+              .map((pubkey) =>
+                  pubkey != null ? Uint8List.fromList(pubkey) : null)
+              .toList()
           : null,
       signatures: input.signatures != null
-          ? input.signatures.map((signature) =>
-              signature != null ? Uint8List.fromList(signature) : null)
+          ? input.signatures!
+              .map((signature) =>
+                  signature != null ? Uint8List.fromList(signature) : null)
+              .toList()
           : null,
     );
   }
@@ -629,11 +638,11 @@ class Input {
 }
 
 class Output {
-  Uint8List script;
-  int value;
-  Uint8List valueBuffer;
-  List<Uint8List> pubkeys;
-  List<Uint8List> signatures;
+  Uint8List? script;
+  int? value;
+  Uint8List? valueBuffer;
+  List<Uint8List?>? pubkeys;
+  List<Uint8List?>? signatures;
 
   Output(
       {this.script,
@@ -641,42 +650,45 @@ class Output {
       this.pubkeys,
       this.signatures,
       this.valueBuffer}) {
-    if (value != null && !isShatoshi(value))
+    if (value != null && !isShatoshi(value!))
       throw ArgumentError('Invalid ouput value');
   }
 
-  factory Output.expandOutput(Uint8List script, [Uint8List ourPubKey]) {
-    if (ourPubKey == null) return new Output();
+  factory Output.expandOutput(Uint8List script, [Uint8List? ourPubKey]) {
+    if (ourPubKey == null) return Output();
     var type = classifyOutput(script);
     if (type == SCRIPT_TYPES['P2WPKH']) {
-      Uint8List wpkh1 =
-          new P2WPKH(data: new PaymentData(output: script)).data.hash;
+      Uint8List wpkh1 = P2WPKH(data: PaymentData(output: script)).data.hash!;
       Uint8List wpkh2 = bcrypto.hash160(ourPubKey);
       if (wpkh1 != wpkh2) throw ArgumentError('Hash mismatch!');
-      return new Output(pubkeys: [ourPubKey], signatures: [null]);
+      return Output(pubkeys: [ourPubKey], signatures: [null]);
     } else if (type == SCRIPT_TYPES['P2PKH']) {
-      Uint8List pkh1 =
-          new P2PKH(data: new PaymentData(output: script)).data.hash;
+      Uint8List pkh1 = P2PKH(data: PaymentData(output: script)).data.hash!;
       Uint8List pkh2 = bcrypto.hash160(ourPubKey);
       if (pkh1 != pkh2) throw ArgumentError('Hash mismatch!');
-      return new Output(pubkeys: [ourPubKey], signatures: [null]);
+      return Output(pubkeys: [ourPubKey], signatures: [null]);
     }
+    throw UnsupportedError('type "$type"');
   }
 
   factory Output.clone(Output output) {
     return new Output(
-      script: output.script != null ? Uint8List.fromList(output.script) : null,
+      script: output.script != null ? Uint8List.fromList(output.script!) : null,
       value: output.value,
       valueBuffer: output.valueBuffer != null
-          ? Uint8List.fromList(output.valueBuffer)
+          ? Uint8List.fromList(output.valueBuffer!)
           : null,
       pubkeys: output.pubkeys != null
-          ? output.pubkeys.map(
-              (pubkey) => pubkey != null ? Uint8List.fromList(pubkey) : null)
+          ? output.pubkeys!
+              .map((pubkey) =>
+                  pubkey != null ? Uint8List.fromList(pubkey) : null)
+              .toList()
           : null,
       signatures: output.signatures != null
-          ? output.signatures.map((signature) =>
-              signature != null ? Uint8List.fromList(signature) : null)
+          ? output.signatures!
+              .map((signature) =>
+                  signature != null ? Uint8List.fromList(signature) : null)
+              .toList()
           : null,
     );
   }
