@@ -1,5 +1,5 @@
-import 'package:bitcoin_flutter/bitcoin_flutter.dart';
 import 'package:bip39/bip39.dart' as bip39;
+import 'package:bitcoin_dart/bitcoin_dart.dart';
 
 main() {
   var seed = bip39.mnemonicToSeed(
