@@ -1,13 +1,15 @@
 import 'dart:typed_data';
-import 'models/networks.dart';
-import 'package:bs58check/bs58check.dart' as bs58check;
+
 import 'package:bech32/bech32.dart';
+import 'package:bs58check/bs58check.dart' as bs58check;
+
+import 'models/networks.dart';
 import 'payments/index.dart' show PaymentData;
 import 'payments/p2pkh.dart';
 import 'payments/p2wpkh.dart';
 
 class Address {
-  static bool validateAddress(String address, [NetworkType nw]) {
+  static bool validateAddress(String address, [NetworkType? nw]) {
     try {
       addressToOutputScript(address, nw);
       return true;
@@ -16,7 +18,7 @@ class Address {
     }
   }
 
-  static Uint8List addressToOutputScript(String address, [NetworkType nw]) {
+  static Uint8List addressToOutputScript(String address, [NetworkType? nw]) {
     NetworkType network = nw ?? bitcoin;
     var decodeBase58;
     var decodeBech32;
@@ -28,7 +30,7 @@ class Address {
         throw new ArgumentError('Invalid version or Network mismatch');
       P2PKH p2pkh =
           new P2PKH(data: new PaymentData(address: address), network: network);
-      return p2pkh.data.output;
+      return p2pkh.data.output!;
     } else {
       try {
         decodeBech32 = segwit.decode(address);
@@ -40,7 +42,7 @@ class Address {
           throw new ArgumentError('Invalid address version');
         P2WPKH p2wpkh = new P2WPKH(
             data: new PaymentData(address: address), network: network);
-        return p2wpkh.data.output;
+        return p2wpkh.data.output!;
       }
     }
     throw new ArgumentError(address + ' has no matching Script');
