@@ -1,8 +1,9 @@
 import 'dart:typed_data';
-import '../src/utils/script.dart' as bscript;
-import 'templates/pubkeyhash.dart' as pubkeyhash;
+
 import 'templates/pubkey.dart' as pubkey;
+import 'templates/pubkeyhash.dart' as pubkeyhash;
 import 'templates/witnesspubkeyhash.dart' as witnessPubKeyHash;
+import 'utils/script.dart' as bscript;
 
 const SCRIPT_TYPES = {
   'P2SM': 'multisig',
@@ -16,7 +17,7 @@ const SCRIPT_TYPES = {
   'WITNESS_COMMITMENT': 'witnesscommitment'
 };
 
-String classifyOutput(Uint8List script) {
+String? classifyOutput(Uint8List script) {
   if (witnessPubKeyHash.outputCheck(script)) return SCRIPT_TYPES['P2WPKH'];
   if (pubkeyhash.outputCheck(script)) return SCRIPT_TYPES['P2PKH'];
   final chunks = bscript.decompile(script);
@@ -24,7 +25,7 @@ String classifyOutput(Uint8List script) {
   return SCRIPT_TYPES['NONSTANDARD'];
 }
 
-String classifyInput(Uint8List script) {
+String? classifyInput(Uint8List script) {
   final chunks = bscript.decompile(script);
   if (chunks == null) throw new ArgumentError('Invalid script');
   if (pubkeyhash.inputCheck(chunks)) return SCRIPT_TYPES['P2PKH'];
@@ -32,7 +33,7 @@ String classifyInput(Uint8List script) {
   return SCRIPT_TYPES['NONSTANDARD'];
 }
 
-String classifyWitness(List<Uint8List> script) {
+String? classifyWitness(List<Uint8List> script) {
   final chunks = bscript.decompile(script);
   if (chunks == null) throw new ArgumentError('Invalid script');
   if (witnessPubKeyHash.inputCheck(chunks)) return SCRIPT_TYPES['P2WPKH'];
