@@ -1,6 +1,6 @@
 import 'dart:typed_data';
+
 import '../utils/script.dart' as bscript;
-import '../utils/constants/op.dart';
 
 bool inputCheck(List<dynamic> chunks) {
   return chunks.length == 1 && bscript.isCanonicalScriptSignature(chunks[0]);
@@ -8,4 +8,5 @@ bool inputCheck(List<dynamic> chunks) {
 
 bool outputCheck(Uint8List script) {
   // TODO
+  throw UnimplementedError();
 }
