@@ -1,7 +1,8 @@
-import 'check_types.dart';
 import 'dart:typed_data';
 
-Uint8List encode(int number, [Uint8List buffer, int offset]) {
+import 'check_types.dart';
+
+Uint8List encode(int number, [Uint8List? buffer, int? offset]) {
   if (!isUint(number, 53)) ;
 
   buffer = buffer ?? new Uint8List(encodingLength(number));
@@ -30,7 +31,7 @@ Uint8List encode(int number, [Uint8List buffer, int offset]) {
   return buffer;
 }
 
-int decode(Uint8List buffer, [int offset]) {
+int decode(Uint8List buffer, [int? offset]) {
   offset = offset ?? 0;
   ByteData bytes = buffer.buffer.asByteData();
   final first = bytes.getUint8(offset);
@@ -59,7 +60,11 @@ int encodingLength(int number) {
   if (!isUint(number, 53)) throw ArgumentError("Expected UInt53");
   return (number < 0xfd
       ? 1
-      : number <= 0xffff ? 3 : number <= 0xffffffff ? 5 : 9);
+      : number <= 0xffff
+          ? 3
+          : number <= 0xffffffff
+              ? 5
+              : 9);
 }
 
 int readUInt64LE(ByteData bytes, int offset) {
