@@ -1,13 +1,15 @@
-import 'package:test/test.dart';
-import 'package:hex/hex.dart';
-import 'dart:typed_data';
-import 'dart:io';
 import 'dart:convert';
-import '../lib/src/ecpair.dart' show ECPair;
-import '../lib/src/models/networks.dart' as NETWORKS;
+import 'dart:io';
+import 'dart:typed_data';
 
-final ONE = HEX
-    .decode('0000000000000000000000000000000000000000000000000000000000000001');
+import 'package:bitcoin_dart/src/ecpair.dart' show ECPair;
+import 'package:bitcoin_dart/src/models/networks.dart' as NETWORKS;
+import 'package:hex/hex.dart';
+import 'package:test/test.dart';
+
+final ONE = HEX.decode(
+        '0000000000000000000000000000000000000000000000000000000000000001')
+    as Uint8List;
 
 main() {
   final fixtures = json.decode(
@@ -29,14 +31,14 @@ main() {
       });
       (fixtures['valid'] as List).forEach((f) {
         test('derives public key for ${f['WIF']}', () {
-          final d = HEX.decode(f['d']);
+          final d = HEX.decode(f['d']) as Uint8List;
           final keyPair = ECPair.fromPrivateKey(d, compressed: f['compressed']);
-          expect(HEX.encode(keyPair.publicKey), f['Q']);
+          expect(HEX.encode(keyPair.publicKey!), f['Q']);
         });
       });
       (fixtures['invalid']['fromPrivateKey'] as List).forEach((f) {
         test('throws ' + f['exception'], () {
-          final d = HEX.decode(f['d']);
+          final d = HEX.decode(f['d']) as Uint8List;
           try {
             expect(ECPair.fromPrivateKey(d), isArgumentError);
           } catch (err) {
@@ -48,7 +50,7 @@ main() {
     group('fromPublicKey', () {
       (fixtures['invalid']['fromPublicKey'] as List).forEach((f) {
         test('throws ' + f['exception'], () {
-          final Q = HEX.decode(f['Q']);
+          final Q = HEX.decode(f['Q']) as Uint8List;
           try {
             expect(ECPair.fromPublicKey(Q), isArgumentError);
           } catch (err) {
@@ -62,7 +64,7 @@ main() {
         test('imports ${f['WIF']}', () {
           final keyPair = ECPair.fromWIF(f['WIF']);
           var network = _getNetwork(f);
-          expect(HEX.encode(keyPair.privateKey), f['d']);
+          expect(HEX.encode(keyPair.privateKey!), f['d']);
           expect(keyPair.compressed, f['compressed']);
           expect(keyPair.network, network);
         });
@@ -121,7 +123,7 @@ main() {
     group('.network', () {
       (fixtures['valid'] as List).forEach((f) {
         test('return ${f['network']} for ${f['WIF']}', () {
-          NETWORKS.NetworkType network = _getNetwork(f);
+          var network = _getNetwork(f);
           final keyPair = ECPair.fromWIF(f['WIF']);
           expect(keyPair.network, network);
         });
@@ -130,14 +132,13 @@ main() {
   });
 }
 
-NETWORKS.NetworkType _getNetwork(f) {
-  var network;
+NETWORKS.NetworkType? _getNetwork(f) {
   if (f['network'] != null) {
     if (f['network'] == 'bitcoin') {
-      network = NETWORKS.bitcoin;
+      return NETWORKS.bitcoin;
     } else if (f['network'] == 'testnet') {
-      network = NETWORKS.testnet;
+      return NETWORKS.testnet;
     }
   }
-  return network;
+  return null;
 }
